requirements.txt
Flask==0.10.1
Flask-OAuthlib==0.9.1
Flask-SQLAlchemy==2.0
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Werkzeug==0.10.4
appnope==0.1.0
decorator==4.0.2
django-dotenv==1.3.0
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
mysqlclient==1.3.6
oauthlib==1.0.3
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
requests==2.7.0
requests-oauthlib==0.5.0
simplegeneric==0.8.1
traitlets==4.0.0
